refactor(browser): split BROWSER_OPERATION_META into a metadata-only module
The browser CLI command needs the operation table to generate its
subcommands, but the table lived in operations.ts next to a module-scope
browserManager import, so registering the command loaded the whole
Playwright stack. The cli/no-daemon-internals allowlist now admits only
the metadata module.
